added logger to unittest
