+++ src/core/index.js
@@ -2,10 +2,12 @@ const DBUtil = require('./utils/dbutil');
 const ObjectNotFound = require('./exceptions/objectnotfound');
 const ErrorCode = require('./enums/errorcode');
 const ValidationError = require('./exceptions/validationerror');
+const logger = require('./logger');
 
 module.exports = {
     DBUtil,
     ObjectNotFound,
     ValidationError,
-    ErrorCode
+    ErrorCode,
+    logger
 }
