+++ src/core/logger.js
@@ -1,0 +1,16 @@
+const winston = require('winston');
+ 
+const logger = winston.createLogger({
+  level: 'info',
+  format: winston.format.json(),
+  defaultMeta: { 
+      server: 'localhost:8010',
+      service: 'ride-service' 
+    },
+  transports: [
+    new winston.transports.File({ filename: 'error.log', level: 'error' }),
+    new winston.transports.File({ filename: 'combined.log' }),
+  ],
+});
+
+module.exports = logger;
